+++ main/__init__.py
@@ -18,10 +18,27 @@ app.app_context().push()
 
 bcrypt=Bcrypt(app)
 login_manager = LoginManager(app) #administra las sesiones en background
-login_manager.login_view = 'login'
+login_manager.login_view = 'users.login'
 login_manager.login_message_category = 'info'
 
 
 
 #pongo las rutas despues de inicializar:
-from main import routes
+from main.users.routes import users_bp
+
+from main.analysis.routes import analysis_bp
+
+from main.candidates.routes import candidates_bp
+
+from main.categories.routes import categories_bp
+
+from main.offers.routes import offers_bp
+
+from main.statuses.routes import statuses_bp
+
+app.register_blueprint(users_bp)
+app.register_blueprint(analysis_bp)
+app.register_blueprint(candidates_bp)
+app.register_blueprint(categories_bp)
+app.register_blueprint(offers_bp)
+app.register_blueprint(statuses_bp)
