+++ main/analysis/analysis.py
@@ -3,39 +3,9 @@ import glob
 import re
 import string
 import pandas as pd
-import numpy as np
 import matplotlib.pyplot as plt
-import seaborn as sns
 
 pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files (x86)\Tesseract-OCR\tesseract.exe'
-
-terms2 = {'Quality/Six Sigma':['black belt','capability analysis','control charts','doe','dmaic','fishbone',
-                              'gage r&r', 'green belt','ishikawa','iso','kaizen','kpi','lean','metrics',
-                              'pdsa','performance improvement','process improvement','quality',
-                              'quality circles','quality tools','root cause','six sigma',
-                              'stability analysis','statistical analysis','tqm'],      
-        'Operations management':['automation','bottleneck','constraints','cycle time','efficiency','fmea',
-                                 'machinery','maintenance','manufacture','line balancing','oee','operations',
-                                 'operations research','optimization','overall equipment effectiveness',
-                                 'pfmea','process','process mapping','production','resources','safety',
-                                 'stoppage','value stream mapping','utilization'],
-        'Supply chain':['abc analysis','apics','customer','customs','delivery','distribution','eoq','epq',
-                        'fleet','forecast','inventory','logistic','materials','outsourcing','procurement',
-                        'reorder point','rout','safety stock','scheduling','shipping','stock','suppliers',
-                        'third party logistics','transport','transportation','traffic','supply chain',
-                        'vendor','warehouse','wip','work in progress'],
-        'Project management':['administration','agile','budget','cost','direction','feasibility analysis',
-                              'finance','kanban','leader','leadership','management','milestones','planning',
-                              'pmi','pmp','problem','project','risk','schedule','scrum','stakeholders'],
-        'Data analytics':['analytics','api','aws','big data','busines intelligence','clustering','code',
-                          'coding','data','database','data mining','data science','deep learning','hadoop',
-                          'hypothesis test','iot','internet','machine learning','modeling','nosql','nlp',
-                          'predictive','programming','python','r','sql','tableau','text mining',
-                          'visualuzation'],
-        'Healthcare':['adverse events','care','clinic','cphq','ergonomics','healthcare',
-                      'health care','health','hospital','human factors','medical','near misses',
-                      'patient','reporting system']}
-
 
 terms = {#'Pensamiento crítico y resolutivo':[],      
         #'Colaboración y trabajo en equipo':[],
