+++ main/main/utils.py
@@ -1,0 +1,6 @@
+NAMETYPE_USER_USER = 'Usuario'
+NAMETYPE_USER_CANDIDATE = 'Candidato'
+NAMETYPE_USER_ADMINISTRATOR = 'Administrador'
+CLASSIFICATION_STATUS_OFFER = 'Trabajo'
+CLASSIFICATION_STATUS_CANDIDATE = 'Candidato'
+
