+++ main/templates/login.html
@@ -41,7 +41,7 @@
         </form>
         <div class="border-top pt-3">
             <small>
-                ¿No tienes cuenta? Regístrate <a class="ml-2" href="{{ url_for('register')}}">aquí</a>
+                ¿No tienes cuenta? Regístrate <a class="ml-2" href="{{ url_for('users.register')}}">aquí</a>
             </small>
         </div>
     </div>
