+++ main/templates/register.html
@@ -91,7 +91,7 @@
         </form>
         <div class="border-top pt-3" >
             <small class="text-muted">
-               ¿Ya tienes cuenta? <a class="ml-2" href="{{ url_for('login') }}">Entra</a>
+               ¿Ya tienes cuenta? <a class="ml-2" href="{{ url_for('users.login') }}">Entra</a>
             </small>
         </div>
     </div>
